refactor(minifier): move code to proper position
