+++ crates/oxc_minifier/src/compressor/fold.rs
@@ -2,13 +2,12 @@
 //!
 //! <https://github.com/google/closure-compiler/blob/master/src/com/google/javascript/jscomp/PeepholeFoldConstants.java>
 
-use oxc_ast::BigUint;
 #[allow(clippy::wildcard_imports)]
 use oxc_hir::hir::*;
-use oxc_hir::hir_util::{IsLiteralValue, MayHaveSideEffects};
+use oxc_hir::hir_util::{get_boolean_value, get_number_value, IsLiteralValue, MayHaveSideEffects};
 use oxc_span::{Atom, Span};
 use oxc_syntax::{
-    operator::{AssignmentOperator, BinaryOperator, LogicalOperator, UnaryOperator},
+    operator::{BinaryOperator, UnaryOperator},
     NumberBase,
 };
 
@@ -201,98 +200,81 @@ impl<'a> Compressor<'a> {
     ) -> Option<Expression<'a>> {
         self.fold_expression(&mut unary_expr.argument);
 
-        let tri_kind = get_boolean_value(&unary_expr.argument);
-
-        if tri_kind == Tri::Unknown {
-            return None;
-        }
-
-        match unary_expr.operator {
-            UnaryOperator::LogicalNot => {
-                if let Expression::NumberLiteral(number_literal) = &unary_expr.argument {
-                    let value = number_literal.value;
-                    if value == 0_f64 || (value - 1_f64).abs() < f64::EPSILON {
-                        None
-                    } else {
-                        // Tri::Unknown has been already filtered out
-                        let bool_literal =
-                            self.hir.boolean_literal(unary_expr.span, tri_kind != Tri::True);
-                        Some(self.hir.literal_boolean_expression(bool_literal))
-                    }
-                } else {
-                    None
+        if let Some(tri_kind) = get_boolean_value(&unary_expr.argument) {
+            match unary_expr.operator {
+                UnaryOperator::LogicalNot => {
+                    if let Expression::NumberLiteral(number_literal) = &unary_expr.argument {
+                        let value = number_literal.value;
+                        if value == 0_f64 || (value - 1_f64).abs() < f64::EPSILON {
+                            return None
+                        } else {
+                            // Tri::Unknown has been already filtered out
+                            let bool_literal =
+                                self.hir.boolean_literal(unary_expr.span, tri_kind);
+                            return Some(self.hir.literal_boolean_expression(bool_literal))
+                        }
+                    } 
                 }
-            }
-            UnaryOperator::UnaryPlus => match &unary_expr.argument {
-                Expression::NumberLiteral(number_literal) => {
-                    let number_literal = self.hir.number_literal(
-                        unary_expr.span,
-                        number_literal.value,
-                        number_literal.raw,
-                        number_literal.base,
-                    );
-                    Some(self.hir.literal_number_expression(number_literal))
-                }
-                Expression::Identifier(ident) => {
-                    if matches!(ident.name.as_str(), "NaN" | "Infinity") {
-                        self.try_detach_unary_op(unary_expr)
-                    } else {
-                        None
-                    }
-                }
-                _ => {
-                    if let Some(value) = get_number_value(&unary_expr.argument) {
-                        let raw = self.hir.new_str(value.to_string().as_str());
+                UnaryOperator::UnaryPlus => match &unary_expr.argument {
+                    Expression::NumberLiteral(number_literal) => {
                         let number_literal = self.hir.number_literal(
                             unary_expr.span,
-                            value,
-                            raw,
-                            NumberBase::Decimal,
-                        );
-                        Some(self.hir.literal_number_expression(number_literal))
-                    } else {
-                        None
-                    }
-                }
-            },
-            UnaryOperator::UnaryNegation => match &unary_expr.argument {
-                Expression::NumberLiteral(number_literal) => {
-                    let value = -number_literal.value;
-                    let raw = self.hir.new_str(value.to_string().as_str());
-                    let number_literal =
-                        self.hir.number_literal(unary_expr.span, value, raw, number_literal.base);
-                    Some(self.hir.literal_number_expression(number_literal))
-                }
-                Expression::Identifier(ident) => {
-                    if ident.name == "NaN" {
-                        self.try_detach_unary_op(unary_expr)
-                    } else {
-                        None
-                    }
-                }
-                _ => None,
-            },
-            UnaryOperator::BitwiseNot => {
-                if let Expression::NumberLiteral(number_literal) = &unary_expr.argument {
-                    let value = number_literal.value;
-                    if value.fract() == 0.0 {
-                        let int_value = ecmascript_to_int32(value);
-                        let number_literal = self.hir.number_literal(
-                            unary_expr.span,
-                            f64::from(!int_value),
+                            number_literal.value,
                             number_literal.raw,
                             number_literal.base,
                         );
-                        Some(self.hir.literal_number_expression(number_literal))
-                    } else {
-                        None
+                        return Some(self.hir.literal_number_expression(number_literal))
                     }
-                } else {
-                    None
+                    Expression::Identifier(ident) => {
+                        if matches!(ident.name.as_str(), "NaN" | "Infinity") {
+                            return self.try_detach_unary_op(unary_expr)
+                        } 
+                    }
+                    _ => {
+                        if let Some(value) = get_number_value(&unary_expr.argument) {
+                            let raw = self.hir.new_str(value.to_string().as_str());
+                            let number_literal = self.hir.number_literal(
+                                unary_expr.span,
+                                value,
+                                raw,
+                                NumberBase::Decimal,
+                            );
+                            return Some(self.hir.literal_number_expression(number_literal))
+                        } 
+                    }
+                },
+                UnaryOperator::UnaryNegation => match &unary_expr.argument {
+                    Expression::NumberLiteral(number_literal) => {
+                        let value = -number_literal.value;
+                        let raw = self.hir.new_str(value.to_string().as_str());
+                        let number_literal =
+                            self.hir.number_literal(unary_expr.span, value, raw, number_literal.base);
+                        return Some(self.hir.literal_number_expression(number_literal))
+                    }
+                    Expression::Identifier(ident) => {
+                        if ident.name == "NaN" {
+                            return self.try_detach_unary_op(unary_expr)
+                        } 
+                    }
+                    _ => {},
+                },
+                UnaryOperator::BitwiseNot => {
+                    if let Expression::NumberLiteral(number_literal) = &unary_expr.argument && number_literal.value.fract() == 0.0 {
+                            let int_value = NumberLiteral::ecmascript_to_int32(number_literal.value);
+                            let number_literal = self.hir.number_literal(
+                                unary_expr.span,
+                                f64::from(!int_value),
+                                number_literal.raw,
+                                number_literal.base,
+                            );
+                            return Some(self.hir.literal_number_expression(number_literal))
+                    } 
                 }
+                _ => {},
             }
-            _ => None,
-        }
+        } 
+
+        None
     }
 
     fn try_detach_unary_op(
@@ -307,154 +289,12 @@ impl<'a> Compressor<'a> {
                         ident.name.clone(),
                         ident.reference_id,
                     );
-                    Some(self.hir.identifier_reference_expression(ident))
-                } else {
-                    None
-                }
+                    return Some(self.hir.identifier_reference_expression(ident))
+                } 
             }
-            _ => None,
+            _ => {},
         }
-    }
-}
 
-/// port from [closure compiler](https://github.com/google/closure-compiler/blob/a4c880032fba961f7a6c06ef99daa3641810bfdd/src/com/google/javascript/jscomp/NodeUtil.java#L348)
-/// Gets the value of a node as a Number, or None if it cannot be converted.
-fn get_number_value(expr: &Expression) -> Option<f64> {
-    match expr {
-        Expression::NumberLiteral(number_literal) => Some(number_literal.value),
-        Expression::UnaryExpression(unary_expr) => match unary_expr.operator {
-            UnaryOperator::UnaryPlus => get_number_value(&unary_expr.argument),
-            UnaryOperator::UnaryNegation => get_number_value(&unary_expr.argument).map(|v| -v),
-            UnaryOperator::BitwiseNot => get_number_value(&unary_expr.argument)
-                .map(|value| f64::from(!ecmascript_to_int32(value))),
-            UnaryOperator::LogicalNot => match get_boolean_value(expr) {
-                Tri::True => Some(1.0),
-                Tri::False => Some(0.0),
-                Tri::Unknown => None,
-            },
-            _ => None,
-        },
-        Expression::BooleanLiteral(bool_literal) => {
-            if bool_literal.value {
-                Some(1.0)
-            } else {
-                Some(0.0)
-            }
-        }
-        Expression::NullLiteral(_) => Some(0.0),
-        _ => None,
-    }
-}
-
-/// port from [closure compiler](https://github.com/google/closure-compiler/blob/a4c880032fba961f7a6c06ef99daa3641810bfdd/src/com/google/javascript/jscomp/base/JSCompDoubles.java#L113)
-/// <https://262.ecma-international.org/5.1/#sec-9.5>
-#[allow(clippy::cast_possible_truncation)] // for `as i32`
-fn ecmascript_to_int32(num: f64) -> i32 {
-    let int32_value = num as i32;
-    if (f64::from(int32_value) - num).abs() < f64::EPSILON {
-        return int32_value;
-    }
-
-    let pos_int = num.signum() * num.abs().floor();
-    let int32bit = pos_int % 2f64.powi(32);
-
-    if int32bit >= 2f64.powi(31) { (int32bit - 2f64.powi(32)) as i32 } else { int32bit as i32 }
-}
-
-/// port from [closure compiler](https://github.com/google/closure-compiler/blob/a4c880032fba961f7a6c06ef99daa3641810bfdd/src/com/google/javascript/jscomp/NodeUtil.java#L109)
-/// Gets the boolean value of a node that represents an expression, or  Tri.UNKNOWN if no
-/// such value can be determined by static analysis.
-/// This method does not consider whether the node may have side-effects.
-fn get_boolean_value(expr: &Expression) -> Tri {
-    let be_tri_boolean = |boolean: bool| {
-        if boolean { Tri::True } else { Tri::False }
-    };
-
-    match expr {
-        Expression::RegExpLiteral(_) | Expression::ArrayExpression(_)| Expression::ArrowExpression(_)| Expression::ClassExpression(_) | Expression::FunctionExpression(_)| Expression::NewExpression(_) |  Expression::ObjectExpression(_) => Tri::True,
-        Expression::NullLiteral(_) => Tri::False,
-        Expression::BooleanLiteral(boolean_literal) => {
-            be_tri_boolean(boolean_literal.value)
-        },
-        Expression::NumberLiteral(number_literal) => {
-            be_tri_boolean(number_literal.value != 0.0)
-        },
-        Expression::BigintLiteral(big_int_literal) => be_tri_boolean(big_int_literal.value == BigUint::default()),
-        Expression::StringLiteral(string_literal) => be_tri_boolean(!string_literal.value.is_empty()),
-        Expression::TemplateLiteral(template_literal) => {
-            if let Some(quasi) = template_literal.quasis.get(0) && quasi.tail {
-                if quasi.value.cooked.as_ref().map_or(false, |cooked| !cooked.is_empty()) {
-                    Tri::True
-                } else {
-                    Tri::False
-                }
-            } else {
-                Tri::Unknown
-            }
-        },
-        Expression::Identifier(ident) => {
-            if expr.is_undefined() {
-                Tri::False
-            } else if  ident.name == "Infinity" {
-                Tri::True
-            } else if ident.name == "NaN" {
-                Tri::False
-            } else {
-                Tri::Unknown
-            }
-        },
-        Expression::AssignmentExpression(assign_expr) => {
-            match assign_expr.operator {
-                AssignmentOperator::LogicalAnd | AssignmentOperator::LogicalOr => {
-                    Tri::Unknown
-                }
-                // For ASSIGN, the value is the value of the RHS.
-                _ =>  get_boolean_value(&assign_expr.right)
-            }
-        },
-        Expression::LogicalExpression(logical_expr) => {
-            match logical_expr.operator {
-                LogicalOperator::And => {
-                    if get_boolean_value(&logical_expr.left) == Tri::True && get_boolean_value(&logical_expr.right) == Tri:: True {
-                        Tri::True
-                    } else {
-                        Tri::False
-                    }
-                },
-                LogicalOperator::Or => {
-                    if get_boolean_value(&logical_expr.left) == Tri::True || get_boolean_value(&logical_expr.right) == Tri:: True {
-                        Tri::True
-                    } else {
-                        Tri::False
-                    }
-                },
-                LogicalOperator::Coalesce => Tri::Unknown
-            }
-        },
-        Expression::SequenceExpression(sequence_expr) => {
-            // For sequence expression, the value is the value of the RHS.
-            sequence_expr.expressions.last().map_or(Tri::Unknown, get_boolean_value)
-        },
-        Expression::UnaryExpression(unary_expr) => {
-            if unary_expr.operator == UnaryOperator::Void {
-                Tri::False
-            } else if matches!(unary_expr.operator, UnaryOperator::BitwiseNot | UnaryOperator::UnaryPlus | UnaryOperator::UnaryNegation) {
-                match &unary_expr.argument {
-                   Expression::NumberLiteral(number_literal) => be_tri_boolean(number_literal.value != 0.0),
-                   Expression::BigintLiteral(big_int_literal) => be_tri_boolean(big_int_literal.value == BigUint::default()),
-                   _ => Tri::Unknown
-                }
-            } else if unary_expr.operator == UnaryOperator::LogicalNot {
-                match get_boolean_value(&unary_expr.argument) {
-                    Tri::True => Tri::False,
-                    Tri::False => Tri::True,
-                    Tri::Unknown => Tri::Unknown
-                }
-            } else {
-                Tri::Unknown
-            }
-        },
-        _ => Tri::Unknown
-
+        None
     }
 }
